requirements.txt
fastapi
uvicorn
sentence-transformers
numpy
requests
dotenv
faiss-cpu